requirements.txt
mysql-connector-python==8.0.33
cryptography==42.0.2